observatory/requirements.txt
django
psycopg2
pyopenssl
pyparsing
appdirs
py-dateutil
django-analytical